VALHALLA: Stub of new "valhalla" engine
This is intended to grow to support the Valhalla trilogy of games by
Vulcan Software, in the form of the "Valhalla Classics" episodic
distributions that target Windows systems.

The Valhalla Classics have the same storyline, puzzles and gameplay as
the Amiga originals but an entirely different set of asset files that
seem aimed at easy consumption using the Windows API and DirectX.

I'm aiming at Valhalla Classics rather than the three Amiga games here
because Valhalla Classics are still available for sale at the time of
writing this, and their distribution files are more accessible than
opaque Amiga floppy disk images.